fix tests

+++ torchax/test/test_interop.py
@@ -5,6 +5,7 @@ import torchax
 from torchax import interop, jax_device
 import torchax
 import jax
+import jax.numpy as jnp
 
 
 def is_tpu_available():
@@ -176,10 +177,10 @@ class InteropTest(unittest.TestCase):
     self.assertEqual(interop.jax_view(dtype), jnp.float32.dtype)
     self.assertEqual(interop.torch_view(interop.jax_view(dtype)), dtype)
     dtype = torch.bfloat16
-    self.assertEqual(interop.jax_view(dtype), jnp.float32.dtype)
+    self.assertEqual(interop.jax_view(dtype), jnp.bfloat16.dtype)
     self.assertEqual(interop.torch_view(interop.jax_view(dtype)), dtype)
     dtype = torch.int32
-    self.assertEqual(interop.jax_view(dtype), jnp.float32.dtype)
+    self.assertEqual(interop.jax_view(dtype), jnp.int32.dtype)
     self.assertEqual(interop.torch_view(interop.jax_view(dtype)), dtype)
